requirements.txt
httpx==0.28.1
beautifulsoup4==4.13.5
lxml==6.0.1
readability-lxml==0.8.4.1
pypdf==6.1.0
langdetect==1.0.9
spacy==3.8.7
tqdm==4.67.1
